add new feature and website version
